requirements.txt
# Core ML and Deep Learning
tensorflow>=2.15.0
keras>=3.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Framework
Flask>=2.3.0
Werkzeug>=2.3.0

# Data Processing and Utilities
joblib>=1.3.0
tqdm>=4.65.0

# Optional: For advanced features
gdown>=4.7.0  # For downloading models from Google Drive
plotly>=5.15.0  # For interactive visualizations
streamlit>=1.25.0  # Alternative web interface

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
